# services/api/openapi.yaml
openapi: 3.1.0
info:
  title: COROC API
  version: 0.2.0
  summary: Gestión y cobro de préstamos personales
  description: |
    Contrato de la API de COROC (§24.1-3 del prompt maestro). Versión 0.2.0: operaciones de la Fase 1 implementadas y verificadas
    contra este contrato en cada prueba (el servidor valida las peticiones con él y las pruebas validan las respuestas).

    **Convenciones**
    - Dinero: enteros en la unidad mínima de la moneda (`COP` = peso, `BRL`/`USD` = centavo). Nunca decimales binarios.
    - Tasas: texto decimal (`"0.02"` equivale a dos por ciento).
    - Fechas civiles `YYYY-MM-DD` en la zona horaria de la empresa; instantes en ISO 8601 con zona.
    - Errores con `application/problem+json` (RFC 9457).
    - Toda petición autenticada opera dentro de la empresa del token; la base de datos aplica Row Level Security.
    - Idempotencia: las operaciones que crean movimientos contables aceptan el encabezado `Idempotency-Key`.
    - Concurrencia optimista: los recursos editables exponen `version`; enviar `If-Match` con ese valor.
    - Tiempo real: `GET /events` (Server-Sent Events) emite `payment.posted`, `intake.updated`, `message.updated`, `dashboard.changed`.
  contact:
    name: COROC
  license:
    name: Propietaria
    identifier: LicenseRef-COROC-Proprietary
servers:
  - url: https://api.coroc.app/v1
    description: Producción
  - url: http://localhost:3000/v1
    description: Desarrollo local
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [users]
      operationId: listUsers
      summary: Usuarios de la empresa
      responses:
        '200':
          description: Usuarios
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
        '4XX': { $ref: '#/components/responses/Problem' }
    post:
      tags: [users]
      operationId: createUser
      summary: Crear usuario (Propietario o Administrador)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, name, role, password]
              properties:
                username: { type: string, pattern: '^[a-z0-9._-]{3,32}$' }
                name: { type: string, minLength: 1, maxLength: 120 }
                role: { $ref: '#/components/schemas/Role' }
                password: { type: string, format: password, minLength: 12, maxLength: 128 }
                email: { type: string, format: email }
                lang: { $ref: '#/components/schemas/Lang' }
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '422': { $ref: '#/components/responses/Problem' }
  /users/{id}:
    patch:
      tags: [users]
      operationId: updateUser
      summary: Activar, desactivar o cambiar rol
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                active: { type: boolean }
                role: { $ref: '#/components/schemas/Role' }
                name: { type: string, minLength: 1, maxLength: 120 }
                email: { type: [string, 'null'], format: email }
      responses:
        '200':
          description: Actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '4XX': { $ref: '#/components/responses/Problem' }
  /users/{id}/sessions:
    delete:
      tags: [users]
      operationId: revokeSessions
      summary: Cerrar remotamente todas las sesiones del usuario
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: Sesiones revocadas }
        '4XX': { $ref: '#/components/responses/Problem' }
...
  /loans/preview:
    post:
      tags: [loans]
      operationId: previewLoan
      summary: Vista previa en vivo del plan de pagos, tasa efectiva anual y control de tope (§8.3)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoanTerms' }
      responses:
        '200':
          description: Plan calculado (no se guarda nada)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoanPreview' }
        '422': { $ref: '#/components/responses/Problem' }
...
  /loans/{id}:
    get:
      tags: [loans]
      operationId: getLoan
      summary: Préstamo con resumen derivado del libro
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Préstamo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Loan' }
        '4XX': { $ref: '#/components/responses/Problem' }
  /loans/{id}/schedule:
    get:
      tags: [loans]
      operationId: getSchedule
      summary: Cuadro de inversión y pagos (§10)
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Cuotas con su estado y soportes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/InstallmentState' }
        '4XX': { $ref: '#/components/responses/Problem' }
  /loans/{id}/ledger:
    get:
      tags: [loans]
      operationId: getLedger
      summary: Libro de movimientos del préstamo
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Movimientos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LedgerEntry' }
        '4XX': { $ref: '#/components/responses/Problem' }
  /loans/{id}/payments:
    post:
      tags: [payments]
      operationId: postPayment
      summary: Registrar un pago manual o en efectivo (§14)
      description: Transacción atómica. Aplica según §9.5, emite el recibo (§15) y encola su envío respetando §11.4.
      parameters: [{ $ref: '#/components/parameters/Id' }, { $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaymentInput' }
      responses:
        '201':
          description: Pago registrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentResult' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /loans/{id}/payments/{entryId}/reversal:
    post:
      tags: [payments]
      operationId: reversePayment
      summary: Reversar un pago (roles Propietario y Administrador)
      description: Crea un contramovimiento con motivo y marca el recibo como ANULADO. Nunca borra.
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { name: entryId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 3 }
      responses:
        '201':
          description: Reverso registrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReversalResult' }
        '409': { $ref: '#/components/responses/Problem' }
  /loans/{id}/receipts:
    get:
      tags: [payments]
      operationId: listReceipts
      summary: Recibos emitidos del préstamo, incluidos los anulados
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Recibos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ReceiptRecord' }
        '4XX': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de acceso de 15 minutos ligado al dispositivo.
    webhookSecret:
      type: apiKey
      in: header
      name: X-COROC-Webhook-Secret
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, maxLength: 128 }
...
  responses:
    Problem:
      description: Error (RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          description: Código estable, p. ej. RATE_CAP_EXCEEDED, DUPLICATE_RECEIPT, CONTACT_RULE_BLOCKED, VERSION_CONFLICT
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    Minor:
      type: integer
      format: int64
      description: Monto en unidades mínimas de la moneda
    Currency: { type: string, enum: [COP, BRL, USD] }
    Lang: { type: string, enum: [es, pt-BR, en] }
    Role: { type: string, enum: [owner, admin, collector, auditor] }
    Frequency: { type: string, enum: [daily, weekly, monthly] }
    Channel: { type: string, enum: [whatsapp, email] }
    DocKind: { type: string, enum: [plan, schedule, receipt_in, receipt_out, statement, payoff, other] }
...
    MessageStatus: { type: string, enum: [ready, scheduled, sent, delivered, read, failed, blocked, cancelled] }
    MessageEvent: { type: string, enum: [welcome, reminder, overdue, receipt, statement, payoff, manual] }
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        name: { type: string }
        email: { type: [string, 'null'] }
        role: { $ref: '#/components/schemas/Role' }
        lang: { $ref: '#/components/schemas/Lang' }
        theme: { type: string, enum: [system, light, dark] }
        active: { type: boolean }
        mfaEnabled: { type: boolean }
        autoLockMinutes: { type: integer }
        lastLoginAt: { type: [string, 'null'], format: date-time }
        permissions: { type: array, items: { type: string } }
...
    LateFeePolicy:
      type: object
      required: [type, value, graceDays]
      properties:
        type: { type: string, enum: [percent_daily, fixed_daily] }
        value: { type: string, description: 'percent_daily: fracción diaria ("0.001"); fixed_daily: unidades mínimas por día' }
        graceDays: { type: integer, minimum: 0 }
    LoanTerms:
      type: object
      required: [principal, currency, method, rate, installments, frequency, disbursementDate]
      properties:
        principal: { type: integer, format: int64, minimum: 1, description: Capital en unidades mínimas }
        currency: { $ref: '#/components/schemas/Currency' }
        method: { type: string, enum: [simple, french] }
        rate: { type: string, pattern: '^[0-9]{1,3}(\.[0-9]{1,10})?$', description: 'simple: tasa sobre el total; french: tasa por período', example: '0.20' }
        installments: { type: integer, minimum: 1, maximum: 3660 }
        frequency: { $ref: '#/components/schemas/Frequency' }
        disbursementDate: { type: string, format: date }
        firstDueDate: { type: string, format: date }
        collectionDays: { type: array, items: { type: integer, minimum: 1, maximum: 7 }, default: [1, 2, 3, 4, 5, 6] }
        excludeHolidays: { type: boolean, default: true }
        monthlyDay: { type: integer, minimum: 1, maximum: 31 }
        roundingUnit: { $ref: '#/components/schemas/Minor' }
...
    ScheduledInstallment:
      type: object
      properties:
        number: { type: integer }
        dueDate: { type: string, format: date }
        amount: { $ref: '#/components/schemas/Minor' }
        principal: { $ref: '#/components/schemas/Minor' }
        interest: { $ref: '#/components/schemas/Minor' }
        balanceAfter: { $ref: '#/components/schemas/Minor' }
    InstallmentState:
      allOf:
        - $ref: '#/components/schemas/ScheduledInstallment'
        - type: object
          properties:
            paid: { $ref: '#/components/schemas/Minor' }
            lateFeeAccrued: { $ref: '#/components/schemas/Minor' }
            lateFeePaid: { $ref: '#/components/schemas/Minor' }
            status: { type: string, enum: [pending, partial, paid, overdue, waived] }
            lastPaymentDate: { type: [string, 'null'], format: date }
            proofDocumentIds: { type: array, items: { type: string, format: uuid } }
            receiptNumbers: { type: array, items: { type: string } }
    RateCapCheck:
      type: object
      properties:
        ok: { type: boolean }
        effectiveAnnual: { type: number }
        cap: { type: [number, 'null'] }
        maxRate: { type: string, description: 'Tasa máxima que cumple, redondeada hacia abajo a una centésima de punto porcentual' }
        missing: { type: boolean, description: 'No hay tope registrado para la fecha; en Colombia impide crear el préstamo' }
    LoanPreview:
      type: object
      properties:
        terms: { $ref: '#/components/schemas/LoanTerms' }
        installments:
          type: array
          items: { $ref: '#/components/schemas/ScheduledInstallment' }
        regularInstallment: { $ref: '#/components/schemas/Minor' }
        totalPayable: { $ref: '#/components/schemas/Minor' }
        totalInterest: { $ref: '#/components/schemas/Minor' }
        effectiveAnnualRate: { type: number }
        rateCap: { $ref: '#/components/schemas/RateCapCheck' }
    NextInstallment:
      type: [object, 'null']
      properties:
        number: { type: integer }
        dueDate: { type: string, format: date }
        outstanding: { $ref: '#/components/schemas/Minor' }
    LoanSummary:
      type: object
      properties:
        totalInstallments: { type: integer }
        paidInstallments: { type: integer }
        remainingInstallments: { type: integer }
        totalPayable: { $ref: '#/components/schemas/Minor' }
        paidTotal: { $ref: '#/components/schemas/Minor' }
        lateFeesOutstanding: { $ref: '#/components/schemas/Minor' }
        balance: { $ref: '#/components/schemas/Minor' }
        next: { $ref: '#/components/schemas/NextInstallment' }
        overdueCount: { type: integer }
        daysPastDue: { type: integer }
        surplus: { $ref: '#/components/schemas/Minor' }
        progress: { type: number, minimum: 0, maximum: 1 }
    Loan:
      type: object
      properties:
        id: { type: string, format: uuid }
        clientId: { type: string, format: uuid }
        contract: { type: string, example: CT-000125 }
        terms: { $ref: '#/components/schemas/LoanTerms' }
        lateFee: { $ref: '#/components/schemas/LateFeePolicy' }
        totalPayable: { $ref: '#/components/schemas/Minor' }
        totalInterest: { $ref: '#/components/schemas/Minor' }
        effectiveAnnualRate: { type: number }
        status: { type: string, enum: [active, closed, written_off] }
        summary: { $ref: '#/components/schemas/LoanSummary' }
        realizedProfit: { $ref: '#/components/schemas/Minor' }
        expectedMethod: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        version: { type: integer }
    AllocationLine:
      type: object
      properties:
        number: { type: integer }
        toLateFee: { $ref: '#/components/schemas/Minor' }
        toInstallment: { $ref: '#/components/schemas/Minor' }
        completed: { type: boolean }
        partial: { type: boolean }
    LedgerEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [disbursement, payment, late_fee, waiver, reversal, adjustment, credit_balance] }
        entryDate: { type: string, format: date }
        recordedAt: { type: string, format: date-time }
        amount: { $ref: '#/components/schemas/Minor' }
        method: { type: string }
        reference: { type: string }
        institution: { type: string }
        source: { type: string }
        documentId: { type: [string, 'null'], format: uuid }
        reversesId: { type: [string, 'null'], format: uuid }
        reversedBy: { type: [string, 'null'], format: uuid }
        reason: { type: string }
        note: { type: string }
        auto: { type: boolean }
        allocation:
          type: array
          items: { $ref: '#/components/schemas/AllocationLine' }
        receiptNumber: { type: [string, 'null'] }
        createdBy: { type: string }
    PaymentInput:
      type: object
      required: [amount, date]
      properties:
        amount: { $ref: '#/components/schemas/Minor' }
        date: { type: string, format: date }
        method: { type: string, maxLength: 60 }
        reference: { type: string, maxLength: 80 }
        institution: { type: string, maxLength: 80 }
        note: { type: string, maxLength: 500 }
        cash: { type: boolean, default: false }
        proofDocumentId: { type: string, format: uuid }
    ReceiptData:
      type: object
      description: Contenido del recibo «Gracias por tu pago» (§15), en el idioma del deudor
      properties:
        lang: { $ref: '#/components/schemas/Lang' }
        number: { type: string }
        issuedAt: { type: string, description: Fecha y hora local YYYY-MM-DD HH:mm }
        company:
          type: object
          properties:
            name: { type: string }
            taxId: { type: string }
            phone: { type: string }
            email: { type: string }
            address: { type: string }
        client:
          type: object
          properties:
            fullName: { type: string }
            code: { type: string }
        contract: { type: string }
        currency: { $ref: '#/components/schemas/Currency' }
        payment:
          type: object
          properties:
            date: { type: string, format: date }
            amount: { $ref: '#/components/schemas/Minor' }
            method: { type: string }
            reference: { type: string }
        totalInstallments: { type: integer }
        coverage: { type: string, example: 'Cuotas 2 y 3 (completas) · Cuota 4 (abono parcial $ 30.000)' }
        remainingInstallments: { type: integer }
        accumulatedPaid: { $ref: '#/components/schemas/Minor' }
        previousBalance: { $ref: '#/components/schemas/Minor' }
        newBalance: { $ref: '#/components/schemas/Minor' }
        next:
          type: object
          properties:
            number: { type: integer }
            dueDate: { type: string, format: date }
            amount: { $ref: '#/components/schemas/Minor' }
        verificationCode: { type: string }
        verificationUrl: { type: string }
        voided: { type: boolean }
    PaymentResult:
      type: object
      required: [entry, receipt, surplus, loanClosed]
      properties:
        entry: { $ref: '#/components/schemas/LedgerEntry' }
        receipt: { $ref: '#/components/schemas/ReceiptData' }
        receiptDocument: { $ref: '#/components/schemas/Document' }
        surplus: { $ref: '#/components/schemas/Minor' }
        message: { $ref: '#/components/schemas/Message' }
        loanClosed: { type: boolean }
    ReversalResult:
      type: object
      required: [entry, balance]
      properties:
        entry: { $ref: '#/components/schemas/LedgerEntry' }
        voidedReceipt: { type: [string, 'null'] }
        balance: { $ref: '#/components/schemas/Minor' }
    ReceiptRecord:
      type: object
      properties:
        number: { type: string }
        entryId: { type: string, format: uuid }
        voided: { type: boolean }
        createdAt: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/ReceiptData' }
    Document:
      type: object
      properties:
        id: { type: string, format: uuid }
        clientId: { type: [string, 'null'], format: uuid }
        loanId: { type: [string, 'null'], format: uuid }
        kind: { $ref: '#/components/schemas/DocKind' }
        name: { type: string }
        mime: { type: string }
        sizeBytes: { type: integer }
        sha256: { type: string }
        version: { type: integer }
        folderPath: { type: string, example: 'Maria Jose Perez Gomez - C000042/CT-000125/03 Recibos emitidos/2026-10-09_1435_RECIBO_RC-000482_CT-000125.pdf' }
        createdAt: { type: string, format: date-time }
...
    ContactDecision:
      type: object
      properties:
        decision: { type: string, enum: [send, reschedule, block] }
        at: { type: string, format: date-time }
        ruleSet: { type: string, example: CO_LEY_2300_2023 }
        reasons:
          type: array
          items:
            type: object
            properties:
              code: { type: string, enum: [OUT_OF_WINDOW, HOLIDAY_SKIPPED, MAX_PER_DAY, CHANNEL_WEEK, TRANSACTIONAL_IMMEDIATE, DEBTOR_EXCEPTION] }
              detail: { type: string }
    Message:
      type: object
      properties:
        id: { type: string, format: uuid }
        clientId: { type: string, format: uuid }
        loanId: { type: string, format: uuid }
        event: { $ref: '#/components/schemas/MessageEvent' }
        kind: { type: string, enum: [collection, transactional] }
        channel: { $ref: '#/components/schemas/Channel' }
        lang: { $ref: '#/components/schemas/Lang' }
        body: { type: string }
        attachmentDocumentId: { type: [string, 'null'], format: uuid }
        status: { $ref: '#/components/schemas/MessageStatus' }
        decision: { $ref: '#/components/schemas/ContactDecision' }
        scheduledAt: { type: [string, 'null'], format: date-time }
        sentAt: { type: [string, 'null'], format: date-time }
        assisted:
          type: object
          description: En modo asistido, enlaces listos para que el usuario envíe
          properties:
            whatsappUrl: { type: string, format: uri }
            mailtoUrl: { type: string }
...